+++ src/pybindings/pybindings_dataset.cpp
@@ -16,7 +16,7 @@ void pybindings_dataset(py::module& root) {
                     py::arg("df"), 
                     py::arg("k") = 10, 
                     py::arg("include_null") = false)
-        .def(py::init<DataFrame, int, int, bool>(), 
+        .def(py::init<DataFrame, int, long unsigned int, bool>(), 
                     py::arg("df"), 
                     py::arg("k") = 10, 
                     py::arg("seed"), 
@@ -37,7 +37,7 @@ void pybindings_dataset(py::module& root) {
                     py::arg("df"), 
                     py::arg("test_ratio") = 0.2, 
                     py::arg("include_null") = false)
-        .def(py::init<const DataFrame&, double, int, bool>(), 
+        .def(py::init<const DataFrame&, double, long unsigned int, bool>(), 
                     py::arg("df"), 
                     py::arg("test_ratio") = 0.2, 
                     py::arg("seed"),
